+++ benches/two_opt.rs
@@ -4,7 +4,6 @@ use rgeometry::data::*;
 
 use num::BigInt;
 use num::BigRational;
-use ordered_float::OrderedFloat;
 
 use rand::Rng;
 use rand::SeedableRng;
@@ -32,8 +31,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     b.iter(|| {
       let mut pts = Vec::new();
       while pts.len() < SET_SIZE {
-        let pt: Point<OrderedFloat<f64>> =
-          rng.sample::<Point<f64, 2>, _>(Standard).map(OrderedFloat);
+        let pt: Point<f64> = rng.sample::<Point<f64, 2>, _>(Standard);
         pts.push(pt)
       }
       two_opt_moves(pts, &mut rng)